+++ lib/libesp32/berry/src/be_module.c
@@ -128,11 +128,15 @@ static char* fixpath(bvm *vm, bstring *path, size_t *size)
     const char *split, *base;
     bvalue *func = vm->cf->func;
     bclosure *cl = var_toobj(func);
+#if BE_DEBUG_SOURCE_FILE
     if (var_isclosure(func)) {
         base = str(cl->proto->source); /* get the source file path */
     } else {
         base = "/";
     }
+#else
+    base = "/";
+#endif
     split = be_splitpath(base);
     *size = split - base + (size_t)str_len(path) + SUFFIX_LEN;
     buffer = be_malloc(vm, *size);